P4ADEV-163: Renamed transactions.tsx for better constistency
