+++ src/App.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next';
 
 import './translations/i18n';
 import { ArcRoutes } from './routes/routes';
-import Transactions from './routes/Transactions';
+import Transaction from './routes/Transaction';
 
 export const App = () => {
   const { t } = useTranslation();
@@ -22,7 +22,7 @@ export const App = () => {
           <Routes>
             <Route path={'/'} element={<>{t('app.title')}</>} />
             <Route path="*" element={<Navigate replace to="/" />} />
-            <Route path={`/${ArcRoutes.TRANSACTION}/:ID`} element={<Transactions />} />
+            <Route path={`/${ArcRoutes.TRANSACTION}/:ID`} element={<Transaction />} />
           </Routes>
         </Layout>
       </BrowserRouter>
